Alterada validação de paradas.*, onde agora são aceitos apenas inteiros, não mais qualquer tipo numérico.

+++ api-teste/app/Http/Requests/LinhaRemoveParadaRequest.php
@@ -50,7 +50,7 @@ class LinhaRemoveParadaRequest extends FormRequest
         $id = $this->segment(4);
         return [
             'paradas' => 'required',
-            'paradas.*' => 'numeric'
+            'paradas.*' => 'integer'
         ];
     }
 }
